move reusable utilities into separate files
