+++ shared/fileParsing/external/yamlFormat/interfaces.ts
@@ -1,0 +1,6 @@
+import { Range } from "../../..";
+
+export interface YamlParsingError {
+    message: string;
+    range: Range;
+}
