+++ shared/fileParsing/external/yamlFormat/parseYamlEnvironmentFile.ts
@@ -1,54 +1,50 @@
-import {
-    isAlias,
-    isCollection,
-    isDocument,
-    isMap,
-    isNode,
-    isScalar,
-    isSeq,
-    LineCounter,
-    parseDocument,
-    YAMLError,
-    YAMLMap,
-    Range as YamlRange,
-    YAMLSeq,
-} from "yaml";
-import { Position, Range, TextDocumentHelper } from "../../..";
+import { LineCounter, parseDocument, YAMLMap } from "yaml";
+import { TextDocumentHelper, YamlParsingError } from "../../..";
+import { getScalarFieldStringValueFromMap } from "../../internal/yamlFormat/yamlMaps/getScalarFieldStringValueFromMap";
+import { mapErrors } from "../../internal/yamlFormat/util/mapErrors";
+import { getTopLevelMapIfExists } from "../../internal/yamlFormat/yamlMaps/getTopLevelMapIfExists";
+import { getYamlSequenceByKeyFromMap } from "../../internal/yamlFormat/yamlMaps/getYamlSequenceByKeyFromMap";
+import { getYamlMapsFromSequence } from "../../internal/yamlFormat/yamlSequences/getYamlMapsFromSequence";
+import { CommonParsingArgs } from "../../internal/yamlFormat/interfaces";
 
-enum EnvironmentTopLevelKey {
+enum EnvironmentKeyName {
     Name = "name",
     Variables = "variables",
 }
 
-export interface YamlParsingError {
-    message: string;
-    range: Range;
+enum VariableProperty {
+    Name = "name",
+    Value = "value",
+    Description = "description",
+    Disabled = "disabled",
+    Secret = "secret",
 }
 
-interface CommonParsingArgs {
-    docHelper: TextDocumentHelper;
-    fullDocumentRange: Range;
+enum VariableValueProperty {
+    Type = "type",
+    Data = "data",
 }
 
-// enum VariableType {
-//     Number = "number",
-//     Boolean = "boolean",
-//     Object = "object",
-// }
+enum VariableType {
+    Number = "number",
+    Boolean = "boolean",
+    Object = "object",
+}
 
-// interface Variable {
-//     name: string;
-//     value?: string | { type: VariableType; data: unknown };
-//     description?: string;
-//     secret: boolean;
-//     disabled: boolean;
-// }
+interface Variable {
+    name: string;
+    value?: string | { type: VariableType; data: string };
+    description?: string;
+    secret: boolean;
+    disabled: boolean;
+}
 
 export function parseYamlEnvironmentFile(docHelper: TextDocumentHelper) {
     const document = parseDocument(docHelper.getText(), {
         lineCounter: new LineCounter(),
     });
     const fullDocumentRange = docHelper.getTextRange();
+    const commonArgs = { docHelper, fullDocumentRange };
 
     if (document.errors.length > 0) {
         console.log("Got technical parsing errors.");
@@ -56,9 +52,8 @@ export function parseYamlEnvironmentFile(docHelper: TextDocumentHelper) {
     }
 
     const maybeTopLevelMap = getTopLevelMapIfExists({
+        ...commonArgs,
         node: document.contents,
-        fullDocumentRange,
-        docHelper,
     });
     if ("error" in maybeTopLevelMap) {
         console.log("Got errors for top level map typeguard.");
@@ -67,32 +62,32 @@ export function parseYamlEnvironmentFile(docHelper: TextDocumentHelper) {
     const { map: topLevelMap } = maybeTopLevelMap;
 
     const maybeNameField = getScalarFieldStringValueFromMap({
+        ...commonArgs,
         map: topLevelMap,
-        key: EnvironmentTopLevelKey.Name,
-        docHelper,
-        fullDocumentRange,
+        key: EnvironmentKeyName.Name,
         isTopLevelMap: true,
     });
-
-    if ("error" in maybeNameField) {
-        return maybeNameField.error;
-    }
-
     const maybeVariablesSequence = getYamlSequenceByKeyFromMap({
+        ...commonArgs,
         map: topLevelMap,
-        key: EnvironmentTopLevelKey.Variables,
-        docHelper,
-        fullDocumentRange,
+        key: EnvironmentKeyName.Variables,
         isTopLevelMap: true,
     });
 
-    if ("error" in maybeVariablesSequence) {
-        return maybeVariablesSequence.error;
+    if ("error" in maybeNameField || "error" in maybeVariablesSequence) {
+        const nameFieldError =
+            "error" in maybeNameField ? maybeNameField.error : undefined;
+        const variablesSeqError =
+            "error" in maybeVariablesSequence
+                ? maybeVariablesSequence.error
+                : undefined;
+        return [nameFieldError, variablesSeqError].filter(
+            (v) => v != undefined,
+        );
     }
 
     const { items: variableItems, errors } = getYamlMapsFromSequence({
-        docHelper,
-        fullDocumentRange,
+        ...commonArgs,
         sequence: maybeVariablesSequence.sequence,
     });
 
@@ -101,174 +96,4 @@ export function parseYamlEnvironmentFile(docHelper: TextDocumentHelper) {
         variables: variableItems,
         errors,
     };
-}
-
-function getTopLevelMapIfExists({
-    node,
-    fullDocumentRange,
-}: CommonParsingArgs & { node: unknown }):
-    { map: YAMLMap<unknown, unknown> } | { error: YamlParsingError } {
-    return isMap(node)
-        ? { map: node }
-        : {
-              error: {
-                  message: `A top level Yaml map is required`,
-                  range: fullDocumentRange,
-              },
-          };
-}
-
-function getScalarFieldStringValueFromMap(
-    args: CommonParsingArgs & {
-        map: YAMLMap<unknown, unknown>;
-        key: string;
-        isTopLevelMap: boolean;
-    },
-): { value: string } | { error: YamlParsingError } {
-    const { map, key, fullDocumentRange, docHelper, isTopLevelMap } = args;
-
-    const existenceError = validateKeyExistsInMap(args);
-    if (existenceError) {
-        return { error: existenceError };
-    }
-
-    const field = map.get(key);
-    return typeof field == "string"
-        ? { value: field }
-        : {
-              error: {
-                  message: `Field '${key}' should be a string.`,
-                  range: isTopLevelMap
-                      ? fullDocumentRange
-                      : ((map.range
-                            ? fromYamlRange(map.range, docHelper)
-                            : fullDocumentRange) ?? fullDocumentRange),
-              },
-          };
-}
-
-function getYamlSequenceByKeyFromMap(
-    args: CommonParsingArgs & {
-        map: YAMLMap<unknown, unknown>;
-        key: string;
-        isTopLevelMap: boolean;
-    },
-): { sequence: YAMLSeq<unknown> } | { error: YamlParsingError } {
-    const { map: parentMap, key, fullDocumentRange, docHelper } = args;
-
-    const existenceError = validateKeyExistsInMap(args);
-    if (existenceError) {
-        return { error: existenceError };
-    }
-
-    const field = parentMap.get(key);
-    return isSeq(field)
-        ? { sequence: field }
-        : {
-              error: {
-                  message: `Field '${key}' should be a Yaml sequence. Got ${JSON.stringify(field)}`,
-                  range:
-                      (parentMap.range
-                          ? fromYamlRange(parentMap.range, docHelper)
-                          : fullDocumentRange) ?? fullDocumentRange,
-              },
-          };
-}
-
-function getYamlMapsFromSequence(
-    args: CommonParsingArgs & {
-        sequence: YAMLSeq<unknown>;
-    },
-): { items: YAMLMap<unknown, unknown>[]; errors: YamlParsingError[] } {
-    const { sequence, fullDocumentRange, docHelper } = args;
-
-    const items: YAMLMap<unknown, unknown>[] = [];
-    const errors: YamlParsingError[] = [];
-
-    for (const item of sequence.items) {
-        if (isMap(item)) {
-            items.push(item);
-            continue;
-        }
-        const range =
-            isScalar(item) ||
-            isSeq(item) ||
-            isCollection(item) ||
-            isAlias(item) ||
-            isDocument(item) ||
-            isNode(item)
-                ? item.range
-                : sequence.range;
-
-        errors.push({
-            message: "Sequence items should be Yaml maps",
-            range:
-                (range ? fromYamlRange(range, docHelper) : fullDocumentRange) ??
-                fullDocumentRange,
-        });
-    }
-
-    return { items, errors };
-}
-
-function validateKeyExistsInMap(
-    args: CommonParsingArgs & {
-        map: YAMLMap<unknown, unknown>;
-        key: string;
-        isTopLevelMap: boolean;
-    },
-): YamlParsingError | undefined {
-    const { map, key, isTopLevelMap, fullDocumentRange, docHelper } = args;
-    return map.has(key)
-        ? undefined
-        : {
-              message: `Mandatory key '${key}' is missing in Yaml map.`,
-              range: isTopLevelMap
-                  ? fullDocumentRange
-                  : ((map.range
-                        ? fromYamlRange(map.range, docHelper)
-                        : fullDocumentRange) ?? fullDocumentRange),
-          };
-}
-
-function fromYamlRange(
-    { "0": startOffset, "2": endOffset }: YamlRange,
-    docHelper: TextDocumentHelper,
-): Range | undefined {
-    const startPosition = docHelper.getPositionForOffset(
-        new Position(0, 0),
-        startOffset,
-    );
-    const endPosition = docHelper.getPositionForOffset(
-        new Position(0, 0),
-        endOffset,
-    );
-
-    return startPosition && endPosition
-        ? new Range(startPosition, endPosition)
-        : undefined;
-}
-
-function mapErrors(
-    errors: YAMLError[],
-    fullContentRange: Range,
-): YamlParsingError[] {
-    return errors.map(({ message, linePos }) => {
-        const startPosition =
-            linePos == undefined
-                ? fullContentRange.start
-                : new Position(linePos[0].line - 1, linePos[0].col - 1);
-        const endPosition =
-            linePos == undefined
-                ? fullContentRange.end
-                : linePos[1]
-                  ? new Position(linePos[1].line - 1, linePos[1].col - 1)
-                  : startPosition;
-
-        return {
-            message,
-            range: new Range(startPosition, endPosition),
-            severity: "ERR",
-        };
-    });
 }
