+++ shared/fileParsing/internal/yamlFormat/interfaces.ts
@@ -1,0 +1,6 @@
+import { Range, TextDocumentHelper } from "../../..";
+
+export interface CommonParsingArgs {
+    docHelper: TextDocumentHelper;
+    fullDocumentRange: Range;
+}
