+++ shared/fileParsing/internal/yamlFormat/yamlMaps/getTopLevelMapIfExists.ts
@@ -1,0 +1,18 @@
+import { isMap, YAMLMap } from "yaml";
+import { YamlParsingError } from "../../../..";
+import { CommonParsingArgs } from "../interfaces";
+
+export function getTopLevelMapIfExists({
+    node,
+    fullDocumentRange,
+}: CommonParsingArgs & { node: unknown }):
+    { map: YAMLMap<unknown, unknown> } | { error: YamlParsingError } {
+    return isMap(node)
+        ? { map: node }
+        : {
+              error: {
+                  message: "A top level Yaml map is required",
+                  range: fullDocumentRange,
+              },
+          };
+}
